Remove duplicate helper file
Change-type: minor
