+++ lib/services/helpers.js
@@ -23,6 +23,12 @@ import skhema from 'skhema'
 import {
 	DetectUA
 } from 'detect-ua'
+import {
+	MESSAGE,
+	WHISPER,
+	SUMMARY,
+	RATING
+} from '../constants'
 
 export const createPermaLink = (card) => {
 	return `${window.location.origin}/${card.id}`
@@ -771,4 +777,18 @@ export const isLocalUrl = (url) => {
 export const toRelativeUrl = (absoluteUrl) => {
 	const url = new URL(absoluteUrl)
 	return `${url.pathname}${url.search}`
+}
+
+export const getTypeBase = (type) => {
+	return type.split('@')[0]
+}
+
+export const isTimelineEvent = (type) => {
+	const typeBase = getTypeBase(type)
+	return _.includes([ MESSAGE, WHISPER, SUMMARY, RATING ], typeBase)
+}
+
+export const isPrivateTimelineEvent = (type) => {
+	const typeBase = getTypeBase(type)
+	return _.includes([ WHISPER, SUMMARY, RATING ], typeBase)
 }
